News:
Habits now save with it's id inside.

Added goals and progress, their models, screens and apiservices.
Goals screen working fine.

Added a new function, creating habits and goals in base of your age, weight or gender on profile creation. To keep basic habits in track.

 Have to add:
The option to "complete" habits, the goals complete on their own.
Graphics on "progress screen".
And a 4th screen which shows profile data and a button to logout.